Content panel and renderer bug fixes

+++ Oxylus/src/Utils/UIUtils.cpp
@@ -29,7 +29,7 @@ namespace Oxylus {
 
   std::string FileDialogs::OpenFile(const std::vector<nfdfilteritem_t>& filter) {
     nfdchar_t* outPath;
-    const nfdresult_t result = NFD_OpenDialog(&outPath, filter.data(), filter.size(), nullptr);
+    const nfdresult_t result = NFD_OpenDialog(&outPath, filter.data(), (uint32_t)filter.size(), nullptr);
     if (result == NFD_OKAY) {
       return outPath;
     }
@@ -38,7 +38,7 @@ namespace Oxylus {
 
   std::string FileDialogs::SaveFile(const std::vector<nfdfilteritem_t>& filter, const char* defaultName) {
     nfdchar_t* outPath;
-    const nfdresult_t result = NFD_SaveDialog(&outPath, filter.data(), filter.size(), nullptr, defaultName);
+    const nfdresult_t result = NFD_SaveDialog(&outPath, filter.data(), (uint32_t)filter.size(), nullptr, defaultName);
     if (result == NFD_OKAY) {
       return outPath;
     }
